+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## v1.1.5
+
+* fixed some url issues in crawler
+* threads added in port scanner
+* fixed status code issue in directory enumeration module
+* more sources added for subdomain enumeration
+    * wayback
+    * sonar
+    * hackertarget
+
+---
+
 ## v1.1.4
 
 * CHANGELOG.md added
@@ -15,5 +27,3 @@
 * port scanner
     * module optimized
 * dedicated wayback module added
-
----